requirements.txt
Appium-Python-Client==0.28
selenium==3.14.0
urllib3==1.23
